Make sure a data connection is available before performing sync tasks
[#48 tagged:inprogress]

+++ src/com/bryanjswift/simplenote/thread/UpdateNoteTask.java
@@ -6,6 +6,7 @@ import android.util.Log;
 
 import com.bryanjswift.simplenote.Constants;
 import com.bryanjswift.simplenote.Preferences;
+import com.bryanjswift.simplenote.manager.Connectivity;
 import com.bryanjswift.simplenote.model.Note;
 import com.bryanjswift.simplenote.net.Api;
 import com.bryanjswift.simplenote.net.ServerCreateCallback;
@@ -34,19 +35,23 @@ public class UpdateNoteTask extends AsyncTask<Note, Void, Void> {
      */
     @Override
     protected Void doInBackground(Note... notes) {
-        final Note note = notes[0];
-        // Note modified, list refreshed by onResume
-        // Send updated note to the server
-        final Api.Credentials credentials = Preferences.getLoginPreferences(context);
-        if (!note.getKey().equals(Constants.DEFAULT_KEY) && note.isDeleted()) {
-            Log.d(LOGGING_TAG, "Deleting note on the server");
-            SimpleNoteApi.delete(note, credentials, new ServerSaveCallback(context, note));
-        } else if (note.getKey().equals(Constants.DEFAULT_KEY)) {
-            Log.d(LOGGING_TAG, "Creating a new note on the server");
-            SimpleNoteApi.create(note, credentials, new ServerCreateCallback(context, note));
+        if (Connectivity.hasInternet(context)) {
+            final Note note = notes[0];
+            // Note modified, list refreshed by onResume
+            // Send updated note to the server
+            final Api.Credentials credentials = Preferences.getLoginPreferences(context);
+            if (!note.getKey().equals(Constants.DEFAULT_KEY) && note.isDeleted()) {
+                Log.d(LOGGING_TAG, "Deleting note on the server");
+                SimpleNoteApi.delete(note, credentials, new ServerSaveCallback(context, note));
+            } else if (note.getKey().equals(Constants.DEFAULT_KEY)) {
+                Log.d(LOGGING_TAG, "Creating a new note on the server");
+                SimpleNoteApi.create(note, credentials, new ServerCreateCallback(context, note));
+            } else {
+                Log.d(LOGGING_TAG, String.format("Sending note '%s' to SimpleNoteApi", note.getKey()));
+                SimpleNoteApi.update(note, credentials, new ServerSaveCallback(context, note));
+            }
         } else {
-            Log.d(LOGGING_TAG, String.format("Sending note '%s' to SimpleNoteApi", note.getKey()));
-            SimpleNoteApi.update(note, credentials, new ServerSaveCallback(context, note));
+            Log.d(LOGGING_TAG, "Unable to send note, no connectivity");
         }
         return null;
     }
